requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.5.2
bcrypt==4.3.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
decorator==5.2.1
Deprecated==1.2.18
ecdsa==0.19.1
exceptiongroup==1.2.2
fabric==3.2.2
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
invoke==2.2.0
Mako==1.3.10
MarkupSafe==2.1.5
paramiko==3.5.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
PyNaCl==1.5.0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.44.0
typing_extensions==4.13.2
uvicorn==0.33.0
wrapt==1.17.2
zipp==3.20.2
